+++ libavcodec/flif16_transform.h
@@ -109,8 +109,8 @@ void ff_flif16_transform_configure(FLIF16TransformContext *, const int);
 
 
 // Maybe put FLIF16Context at first in list
-int8_t ff_flif16_transform_read(FLIF16TransformContext *, FLIF16Context *,
-                                 FLIF16RangesContext *);
+int ff_flif16_transform_read(FLIF16TransformContext *, FLIF16Context *,
+                             FLIF16RangesContext *);
 
 FLIF16RangesContext* ff_flif16_transform_meta(FLIF16Context *,
                                               FLIF16PixelData *,
@@ -118,8 +118,8 @@ FLIF16RangesContext* ff_flif16_transform_meta(FLIF16Context *,
                                               FLIF16TransformContext *, 
                                               FLIF16RangesContext *);
 
-int8_t ff_flif16_transform_reverse(FLIF16Context *, FLIF16TransformContext*, FLIF16PixelData*,
-                                    uint8_t, uint8_t);
+int ff_flif16_transform_reverse(FLIF16Context *, FLIF16TransformContext*, FLIF16PixelData*,
+                                uint8_t, uint8_t);
 
 void ff_flif16_transforms_close(FLIF16TransformContext *);
